Update GLProgram/Shader code

+++ OpenGL/src/tests/TestTexture2D.cpp
@@ -25,7 +25,7 @@ namespace test
 
 		m_VAO = std::make_unique<VertexArray>();
 
-		m_VertexBuffer = std::make_unique<VertexBuffer>(positions, sizeof(positions));
+		m_VertexBuffer = std::make_unique<VertexBuffer>(positions, static_cast<unsigned int>(sizeof(positions)));
 		VertexBufferLayout layout;
 		layout.Push<float>(2);
 		layout.Push<float>(2);
